* http://code.google.com/p/lispbuilder/issues/detail?id=13
It is now possible to get at the the SDL_Event in the WITH-EVENTS macro 
using SDL-EVENT within the scope of WITH-EVENTS. 
It is also possible to create a new SDL_Event and assign this to WITH-EVENTS 
using a new &optional sdl-event parameter.

Use sdl:*sdl-event* to access the current SDL_Event foreign object.



git-svn-id: http://lispbuilder.googlecode.com/svn/trunk@1467 ad8a5436-b7f1-11dd-b55b-2da8d274f894